requirements.txt
psycopg2-binary
pydantic
typer==0.5.0
